fix: la-135 aplit types
